Tickets and Order are completed.

+++ src/main/java/com/demo/bean/OrdersBean.java
@@ -37,7 +37,7 @@ public class OrdersBean {
 	private String stockType;
 	private String technician;
 	private String approver;
-	private Integer recordID;
+	private Long recordID;
 	
 	
 }
