+++ src/main/java/com/demo/dao/OrderDetailsDaoInt.java
@@ -7,8 +7,8 @@ import com.demo.model.OrderDetails;
 public interface OrderDetailsDaoInt {
 	
 	String saveOrderDetails(List<OrderDetails> orderDetails);
-	List<OrderDetails> getOrderDetailsByOrderNum(Integer recordID);
-	List<OrderDetails> getOrderDetailsByOrderNum(String key,Integer recordID);
+	List<OrderDetails> getOrderDetailsByOrderNum(Long recordID);
+	List<OrderDetails> getOrderDetailsByOrderNum(String key,Long recordID);
 	List<OrderDetails> getAllOrderDetails();
 	List<OrderDetails> getAllAvailableOrderDetails(String technician);
 	List<OrderDetails> getOrderDetailsByTechnician(String email);
