+++ src/main/java/com/demo/dao/OrderHistoryDaoInt.java
@@ -8,8 +8,8 @@ import com.demo.model.OrderHistory;
 public interface OrderHistoryDaoInt {
 	
 	void insetOrderHistory(OrderHeader order);
-	List<OrderHistory> getAllOrderHistoryByOrderNumber(int recordID);
+	List<OrderHistory> getAllOrderHistoryByOrderNumber(Long recordID);
 	List<OrderHistory> getAllOrderHistoryByOrderNumber();
-	List<OrderHistory> getAllOrderHistoryTicketNumber(int ticketNumber);
+	List<OrderHistory> getAllOrderHistoryTicketNumber(Long ticketNumber);
 
 }
