+++ src/main/java/com/demo/service/OrderDeliveryServiceInt.java
@@ -14,5 +14,5 @@ import com.itextpdf.text.zugferd.exceptions.DataIncompleteException;
 import com.itextpdf.text.zugferd.exceptions.InvalidCodeException;
 import com.itextpdf.xmp.XMPException;
 public interface OrderDeliveryServiceInt {
-	void createPdf(Integer recordID)throws ParserConfigurationException, SAXException, TransformerException, IOException, DocumentException, XMPException, ParseException, DataIncompleteException, InvalidCodeException;
+	void createPdf(Long recordID)throws ParserConfigurationException, SAXException, TransformerException, IOException, DocumentException, XMPException, ParseException, DataIncompleteException, InvalidCodeException;
 }
