+++ src/main/java/com/demo/service/OrderDetailsInt.java
@@ -5,11 +5,11 @@ import java.util.List;
 import com.demo.model.OrderDetails;
 
 public interface OrderDetailsInt {
-	List<OrderDetails> getOrderDetailsByOrderNum(Integer recordID);
+	List<OrderDetails> getOrderDetailsByOrderNum(Long recordID);
 	List<OrderDetails> getAllOrderDetails();
 	List<OrderDetails> getOrderDetailsByTechnician(String email);
 	List<OrderDetails> getAllAvailableOrderDetails(String technician);
-	List<OrderDetails> getOrderDetailsByOrderNum(String key,Integer recordID);
+	List<OrderDetails> getOrderDetailsByOrderNum(String key,Long recordID);
 	List<OrderDetails> getAllBootStockOrders();
 	List<OrderDetails> getAllSiteStockOrders();
 	List<OrderDetails> getAllAvailableOrderDetailsForCustomer(String customerName);
