+++ src/main/java/com/demo/service/OrderHistoryInt.java
@@ -5,7 +5,7 @@ import java.util.List;
 import com.demo.model.OrderHistory;
 
 public interface OrderHistoryInt {
-	List<OrderHistory> getAllOrderHistoryByOrderNumber(int recordID);
-	List<OrderHistory> getAllOrderHistoryTicketNumber(int ticketNumber);
+	List<OrderHistory> getAllOrderHistoryByOrderNumber(Long recordID);
+	List<OrderHistory> getAllOrderHistoryTicketNumber(Long ticketNumber);
 
 }
