+++ src/main/java/com/demo/service/TicketHistoryInt.java
@@ -5,6 +5,6 @@ import java.util.List;
 import com.demo.model.TicketHistory;
 
 public interface TicketHistoryInt {
-	List<TicketHistory> getHistoryByTicketNumber(int ticketNumber);
+	List<TicketHistory> getHistoryByTicketNumber(Long ticketNumber);
 
 }
