+++ src/main/java/com/demo/service/impl/OrderDeliveryService.java
@@ -21,18 +21,7 @@ import com.itextpdf.text.DocumentException;
 import com.itextpdf.text.zugferd.exceptions.DataIncompleteException;
 import com.itextpdf.text.zugferd.exceptions.InvalidCodeException;
 import com.itextpdf.xmp.XMPException;
-import java.io.IOException;
-import java.text.ParseException;
 
-import javax.xml.parsers.ParserConfigurationException;
-import javax.xml.transform.TransformerException;
-
-import org.xml.sax.SAXException;
-
-import com.itextpdf.text.DocumentException;
-import com.itextpdf.text.zugferd.exceptions.DataIncompleteException;
-import com.itextpdf.text.zugferd.exceptions.InvalidCodeException;
-import com.itextpdf.xmp.XMPException;
 
 @Service("orderDelivery")
 @Transactional
@@ -42,7 +31,7 @@ public class OrderDeliveryService implements OrderDeliveryServiceInt{
 	@Autowired
 	private OrderReportsDaoInt reportsDaoInt;
 	@Override
-	public void createPdf(Integer recordID) throws ParserConfigurationException, SAXException, TransformerException, IOException, DocumentException, XMPException, ParseException, DataIncompleteException, InvalidCodeException{
+	public void createPdf(Long recordID) throws ParserConfigurationException, SAXException, TransformerException, IOException, DocumentException, XMPException, ParseException, DataIncompleteException, InvalidCodeException{
 		reportsDaoInt.createPdf(recordID);
 		
 	}
